15_screen_recorder/requirements.txt
opencv-python
numpy
pyautogui
pillow
opencv-python-headless>=4.5.0 